update scoring to be from 0-1

+++ src/tavi/library/storage/loader/default_loader.py
@@ -23,7 +23,7 @@ class DefaultLoader(AbstractLoader):
         """Get scan type (NONE for default loader)."""
         return RawScanType.NONE
 
-    def get_score(self, path: str) -> int:
+    def get_score(self, path: str) -> float:
         """Get score for scan."""
         return 0
